refactor(state): split narrativeStore into per-concern modules
narrativeStore.ts was a 1,207-line file mixing segment CRUD, decision
handling, world-state impact extraction, ending synthesis, and the persist
config. Following the loreStore split precedent:

- narrativeStore.types.ts: the store interface + factory set/get shapes
  (lives apart so action factories don't import the store module back)
- narrativeStore.state.ts: initial state
- narrativeStore.worldStateImpacts.ts: pure impact extraction
  (extractWorldStateImpacts, extractDecisionContext, event patterns)
- narrativeStore.segments.ts / .decisions.ts / .endings.ts: action factories
- narrativeStore.persistence.ts: persist options, moved verbatim (same store
  name/version/serialization, so existing player data is untouched)
- narrativeStore.ts: ~45-line composition

createIndexedDBStorage gains a generic param (default unknown) so the
explicitly-typed persist options compile; inline-options stores unaffected.
Also refreshed the stale manual mock (endedSessions, hydration flag, ending
setters were missing).
